Added test tiddler and loading of SyntaxHighlighterPlugin.js
The MarkupPostBody tiddler was changed to load the loadExternal.js script,
which will load external scripts found in the ExternalScripts tiddler. (As
suggested by one of the canonical ways of plugin development.)

The StyleSheet tiddler was modified to include the ShCore.css and
ShThemeDefault.css tiddlers, used for syntax highlighting.

A 'Test SyntaxHighlighter3Plugin' tiddler was added to test
syntax higlighting.